resources/requirements.txt
Flask==3.1.2
beautifulsoup4==4.14.3
bs4==0.0.2
frozendict==2.4.7
geopy==2.4.1
ipython==8.18.1
kafka-python==2.3.0
matplotlib==3.9.4
py4j==0.10.9.9
pyelasticsearch==1.4.1
pymongo==4.15.5
requests==2.32.5
scipy==1.13.1
selenium==3.141.0
scikit-learn==1.6.1
tabulate==0.9.0
tldextract==5.3.0
wikipedia==1.4.0
iso8601==2.1.0
urllib3==1.24
joblib==1.5.3
bidict==0.23.1
eventlet==0.40.4
flask-socketio==5.6.0
greenlet==3.2.4
h11==0.16.0
python-engineio==4.13.0
python-socketio==5.16.0
simple-websocket==1.1.0
wsproto==1.2.0
